add default network and supported networks to environment variables
