+++ packages/dapp/.env
@@ -2,4 +2,5 @@ REACT_APP_MAINNET_GRAPHQL_URI="https://graphql.civil.co/v1/query"
 REACT_APP_RINKEBY_GRAPHQL_URI="https://graphql.staging.civil.app/v1/query"
 REACT_APP_GANANCHE_GRAPHQL_URI="http://localhost:8080/v1/query"
 REACT_APP_DISCOURSE_URL="https://discourse.staging.civil.app"
-REACT_APP_DEFAULT_NETWORK="4"
+REACT_APP_DEFAULT_ETHEREUM_NETWORK="4"
+REACT_APP_SUPPORTED_NETWORKS="1,4,50
