+++ packages/dapp/src/helpers/config.ts
@@ -3,6 +3,8 @@ export interface ConfigType {
   RINKEBY_GRAPHQL_URI?: string;
   GANANCHE_GRAPHQL_URI?: string;
   DISCOURSE_URL?: string;
+  DEFAULT_ETHEREUM_NETWORK?: string;
+  SUPPORTED_NETWORKS?: string;
 }
 
 const defaultConfig: ConfigType = {
@@ -10,6 +12,8 @@ const defaultConfig: ConfigType = {
   RINKEBY_GRAPHQL_URI: process.env.REACT_APP_RINKEBY_GRAPHQL_URI,
   GANANCHE_GRAPHQL_URI: process.env.REACT_APP_GANANCHE_GRAPHQL_URI,
   DISCOURSE_URL: process.env.REACT_APP_DISCOURSE_URL,
+  DEFAULT_ETHEREUM_NETWORK: process.env.REACT_APP_DEFAULT_ETHEREUM_NETWORK,
+  SUPPORTED_NETWORKS: process.env.REACT_APP_SUPPORTED_NETWORKS,
 };
 
 const serverConfigString = (window as any).SERVER_CONFIG;
