+++ packages/dapp/src/redux/reducers/network.ts
@@ -1,7 +1,8 @@
 import { AnyAction } from "redux";
 import { networkActions } from "../actionCreators/network";
+import config from "../../helpers/config";
 
-export function network(state: string = "4", action: AnyAction): string {
+export function network(state: string = config.DEFAULT_ETHEREUM_NETWORK || "3", action: AnyAction): string {
   switch (action.type) {
     case networkActions.SET_NETWORK:
       return action.data;
@@ -10,7 +11,7 @@ export function network(state: string = "4", action: AnyAction): string {
   }
 }
 
-export function networkName(state: string = "rinkeby", action: AnyAction): string {
+export function networkName(state: string = "", action: AnyAction): string {
   switch (action.type) {
     case networkActions.SET_NETWORK_NAME:
       return action.data;
