+++ packages/utils/src/civilHelpers.ts
@@ -101,32 +101,6 @@ export enum GovernmentParameters {
   govtPRevealStageLen = "govtPRevealStageLen",
 }
 
-export const supportedNetworks: number[] = [
-  1, // mainnet
-  4, // rinkeby
-  50, // ganache
-];
-
-export const formattedNetworkNames: { [index: string]: string } = {
-  1: "Main Ethereum Network",
-  4: "Rinkeby Test Network",
-  50: "Localhost 8545",
-};
-
-export function getFormattedNetworkNames(networkIDs: number[]): string[] {
-  return networkIDs.map(id => {
-    return formattedNetworkNames[id.toString()];
-  });
-}
-
-export function isNetworkSupported(network: string | number): boolean {
-  let networkKey = network;
-  if (typeof networkKey === "string") {
-    networkKey = parseInt(networkKey, 10);
-  }
-  return supportedNetworks.includes(networkKey);
-}
-
 export function getInfuraUrlFromIpfs(ipfsUrl: string): string {
   const hash = ipfsUrl.split("://")[1];
   return `https://ipfs.infura.io/ipfs/${hash}`;
